+++ FrugalForge/FrugalForge_Data_Producers.lua
@@ -63,4 +63,102 @@
       { itemId = 10620, qty = 1 },
     },
   },
+  {
+    outputItemId = 10939, -- Greater Magic Essence
+    outputQty = 1,
+    reagents = {
+      { itemId = 10938, qty = 3 }, -- Lesser Magic Essence
+    },
+  },
+  {
+    outputItemId = 11082, -- Greater Astral Essence
+    outputQty = 1,
+    reagents = {
+      { itemId = 10998, qty = 3 }, -- Lesser Astral Essence
+    },
+  },
+  {
+    outputItemId = 11135, -- Greater Mystic Essence
+    outputQty = 1,
+    reagents = {
+      { itemId = 11134, qty = 3 }, -- Lesser Mystic Essence
+    },
+  },
+  {
+    outputItemId = 11175, -- Greater Nether Essence
+    outputQty = 1,
+    reagents = {
+      { itemId = 11174, qty = 3 }, -- Lesser Nether Essence
+    },
+  },
+  {
+    outputItemId = 16203, -- Greater Eternal Essence
+    outputQty = 1,
+    reagents = {
+      { itemId = 16202, qty = 3 }, -- Lesser Eternal Essence
+    },
+  },
+  {
+    outputItemId = 22446, -- Greater Planar Essence
+    outputQty = 1,
+    reagents = {
+      { itemId = 22447, qty = 3 }, -- Lesser Planar Essence
+    },
+  },
+  {
+    outputItemId = 34055, -- Greater Cosmic Essence
+    outputQty = 1,
+    reagents = {
+      { itemId = 34056, qty = 3 }, -- Lesser Cosmic Essence
+    },
+  },
+  {
+    outputItemId = 10938, -- Lesser Magic Essence
+    outputQty = 3,
+    reagents = {
+      { itemId = 10939, qty = 1 }, -- Greater Magic Essence
+    },
+  },
+  {
+    outputItemId = 10998, -- Lesser Astral Essence
+    outputQty = 3,
+    reagents = {
+      { itemId = 11082, qty = 1 }, -- Greater Astral Essence
+    },
+  },
+  {
+    outputItemId = 11134, -- Lesser Mystic Essence
+    outputQty = 3,
+    reagents = {
+      { itemId = 11135, qty = 1 }, -- Greater Mystic Essence
+    },
+  },
+  {
+    outputItemId = 11174, -- Lesser Nether Essence
+    outputQty = 3,
+    reagents = {
+      { itemId = 11175, qty = 1 }, -- Greater Nether Essence
+    },
+  },
+  {
+    outputItemId = 16202, -- Lesser Eternal Essence
+    outputQty = 3,
+    reagents = {
+      { itemId = 16203, qty = 1 }, -- Greater Eternal Essence
+    },
+  },
+  {
+    outputItemId = 22447, -- Lesser Planar Essence
+    outputQty = 3,
+    reagents = {
+      { itemId = 22446, qty = 1 }, -- Greater Planar Essence
+    },
+  },
+  {
+    outputItemId = 34056, -- Lesser Cosmic Essence
+    outputQty = 3,
+    reagents = {
+      { itemId = 34055, qty = 1 }, -- Greater Cosmic Essence
+    },
+  },
 }